+++ curator-x-async/src/main/java/org/apache/curator/x/async/api/AsyncRemoveWatchesBuilder.java
@@ -16,9 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.curator.x.async;
+package org.apache.curator.x.async.api;
 
 import org.apache.curator.framework.api.CuratorWatcher;
+import org.apache.curator.x.async.AsyncStage;
 import org.apache.zookeeper.Watcher;
 import java.util.Set;
