+++ curator-x-async/src/main/java/org/apache/curator/x/async/api/AsyncSyncBuilder.java
@@ -16,21 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+package org.apache.curator.x.async.api;
 
-package org.apache.curator.x.async;
+import org.apache.curator.x.async.AsyncStage;
 
 /**
- * Terminal operation for various builders
+ * Builder for syncs
  */
-public interface AsyncPathAndBytesable<T> extends AsyncPathable<T>
+public interface AsyncSyncBuilder extends
+    AsyncPathable<AsyncStage<Void>>
 {
-    /**
-     * Commit the currently building operation using the given path and data
-     * and invoke ZooKeeper
-     *
-     * @param path the path
-     * @param data the data
-     * @return usually an async stage
-     */
-    T forPath(String path, byte[] data);
 }
