+++ curator-x-async/src/main/java/org/apache/curator/x/async/api/WatchedAsyncCuratorFramework.java
@@ -16,7 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.curator.x.async;
+package org.apache.curator.x.async.api;
+
+import org.apache.curator.x.async.api.AsyncExistsBuilder;
+import org.apache.curator.x.async.api.AsyncGetChildrenBuilder;
+import org.apache.curator.x.async.api.AsyncGetConfigBuilder;
+import org.apache.curator.x.async.api.AsyncGetDataBuilder;
 
 /**
  * operations that support watching
