+++ curator-x-async/src/main/java/org/apache/curator/x/async/details/AsyncCreateBuilderImpl.java
@@ -21,10 +21,10 @@ package org.apache.curator.x.async.details;
 import org.apache.curator.framework.api.UnhandledErrorListener;
 import org.apache.curator.framework.imps.CreateBuilderImpl;
 import org.apache.curator.framework.imps.CuratorFrameworkImpl;
-import org.apache.curator.x.async.AsyncCreateBuilder;
-import org.apache.curator.x.async.AsyncPathAndBytesable;
+import org.apache.curator.x.async.api.AsyncCreateBuilder;
+import org.apache.curator.x.async.api.AsyncPathAndBytesable;
 import org.apache.curator.x.async.AsyncStage;
-import org.apache.curator.x.async.CreateOption;
+import org.apache.curator.x.async.api.CreateOption;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.data.ACL;
 import org.apache.zookeeper.data.Stat;
