+++ curator-x-async/src/main/java/org/apache/curator/x/async/details/AsyncRemoveWatchesBuilderImpl.java
@@ -22,10 +22,10 @@ import org.apache.curator.framework.api.CuratorWatcher;
 import org.apache.curator.framework.api.UnhandledErrorListener;
 import org.apache.curator.framework.imps.CuratorFrameworkImpl;
 import org.apache.curator.framework.imps.RemoveWatchesBuilderImpl;
-import org.apache.curator.x.async.AsyncPathable;
-import org.apache.curator.x.async.AsyncRemoveWatchesBuilder;
+import org.apache.curator.x.async.api.AsyncPathable;
+import org.apache.curator.x.async.api.AsyncRemoveWatchesBuilder;
 import org.apache.curator.x.async.AsyncStage;
-import org.apache.curator.x.async.RemoveWatcherOption;
+import org.apache.curator.x.async.api.RemoveWatcherOption;
 import org.apache.zookeeper.Watcher;
 import java.util.Collections;
 import java.util.Objects;
