+++ curator-x-async/src/main/java/org/apache/curator/x/async/details/AsyncTransactionOpImpl.java
@@ -25,13 +25,13 @@ import org.apache.curator.framework.api.transaction.CuratorOp;
 import org.apache.curator.framework.api.transaction.TransactionCreateBuilder;
 import org.apache.curator.framework.api.transaction.TransactionSetDataBuilder;
 import org.apache.curator.framework.imps.CuratorFrameworkImpl;
-import org.apache.curator.x.async.AsyncPathAndBytesable;
-import org.apache.curator.x.async.AsyncPathable;
-import org.apache.curator.x.async.AsyncTransactionCheckBuilder;
-import org.apache.curator.x.async.AsyncTransactionCreateBuilder;
-import org.apache.curator.x.async.AsyncTransactionDeleteBuilder;
-import org.apache.curator.x.async.AsyncTransactionOp;
-import org.apache.curator.x.async.AsyncTransactionSetDataBuilder;
+import org.apache.curator.x.async.api.AsyncPathAndBytesable;
+import org.apache.curator.x.async.api.AsyncPathable;
+import org.apache.curator.x.async.api.AsyncTransactionCheckBuilder;
+import org.apache.curator.x.async.api.AsyncTransactionCreateBuilder;
+import org.apache.curator.x.async.api.AsyncTransactionDeleteBuilder;
+import org.apache.curator.x.async.api.AsyncTransactionOp;
+import org.apache.curator.x.async.api.AsyncTransactionSetDataBuilder;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.data.ACL;
 import java.util.List;
